+++ src/components/base/Button/index.jsx
@@ -1,6 +1,4 @@
-import { COLOR } from '@/styles/color';
-import styled from '@emotion/styled';
-import { forwardRef } from 'react';
+import { StyledButton } from './styles';
 
 function Button({ children, as, bgcolor, color, round, disabled, onClick, ...props }) {
   return (
@@ -19,17 +17,3 @@ function Button({ children, as, bgcolor, color, round, disabled, onClick, ...pro
 }
 
 export default Button;
-
-const View = forwardRef(({ as, style, ...props }, ref) => {
-  const Element = as || 'button';
-  return <Element ref={ref} style={{ ...style }} {...props} />;
-});
-View.displayName = 'View';
-
-const StyledButton = styled(View)`
-  background-color: ${({ bgcolor }) => (bgcolor ? bgcolor : COLOR.GRAY)};
-  color: ${({ color }) => (color ? color : COLOR.DARK)};
-  border-radius: ${({ round }) => (round ? '0.3rem' : 'none')};
-  border: none;
-  cursor: ${({ disabled }) => (disabled ? 'not-allowed' : 'pointer')};
-`;
